move passport-setup to a dedicated folder.
